GET_ALL Api , initial data load, PersonDetailsEntity mapper to PersonDetailsModel completed

+++ sparknetworks-filtermatches-backend/src/main/java/com/sparknetworks/backend/repositories/PersonDetailsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sparknetworks.backend.entities.PersonDetailsEntity;
 
-public interface PersonDetailsRepository extends JpaRepository<PersonDetailsEntity, Long> {
+public interface PersonDetailsRepository extends JpaRepository<PersonDetailsEntity, Integer> {
 
 }
